Revert "Add some casts for %j"
This reverts commit c4d2e483a39176a476c56d35879423fe6e33c0cd.

(Merged from https://github.com/openssl/openssl/pull/4160)

+++ crypto/asn1/x_int64.c
@@ -102,8 +102,8 @@ static int uint64_print(BIO *out, ASN1_VALUE **pval, const ASN1_ITEM *it,
                         int indent, const ASN1_PCTX *pctx)
 {
     if ((it->size & INTxx_FLAG_SIGNED) == INTxx_FLAG_SIGNED)
-        return BIO_printf(out, "%jd\n", (intmax_t)**(int64_t **)pval);
-    return BIO_printf(out, "%ju\n", (uintmax_t)**(uint64_t **)pval);
+        return BIO_printf(out, "%jd\n", **(int64_t **)pval);
+    return BIO_printf(out, "%ju\n", **(uint64_t **)pval);
 }
 
 /* 32-bit variants */
